+++ src/pages/CardList/index.jsx
@@ -3,11 +3,12 @@ import { usePageContext } from 'contexts/PageContext';
 import { useCardDataContext } from 'contexts/CardDataContext';
 import useModal from 'hooks/useModal';
 
-import Button from 'components/@common/Button';
 import { Card } from 'components';
 
-import { CARD_EDITOR_MODE, PAGE_LIST } from 'constants/';
+import { CARD_EDIT_TARGET_INDEX, PAGE_LIST } from 'constants/';
 import { CardWallet, ButtonAddCard } from './styles';
+
+import CardEditConfirm from './Containers/CardEditConfirm';
 
 function CardList() {
   const { setPageTitle, setPageLocation } = usePageContext();
@@ -21,7 +22,7 @@ function CardList() {
     handleModalClose: handleEditModalClose,
   } = useModal();
 
-  const { cardList, setCardEditIndex, removeCardData } = useCardDataContext();
+  const { cardList, setCardEditIndex } = useCardDataContext();
 
   const onClickCard = (index) => {
     setCardFocus(index);
@@ -29,26 +30,8 @@ function CardList() {
   };
 
   const onClickAddButton = () => {
-    setCardEditIndex(CARD_EDITOR_MODE.NEW);
+    setCardEditIndex(CARD_EDIT_TARGET_INDEX.NEW);
     setPageLocation(PAGE_LIST.CARD_EDITOR);
-  };
-
-  const onClickEditButton = () => {
-    setCardEditIndex(focusCardIndex);
-    setPageLocation(PAGE_LIST.CARD_EDITOR);
-  };
-
-  const onClickDeleteButton = async () => {
-    if (!confirm('정말 해당 카드를 제거하시겠습니까?')) return;
-
-    try {
-      await removeCardData(focusCardIndex);
-    } catch (error) {
-      alert(error.message);
-      return;
-    }
-
-    handleEditModalClose();
   };
 
   return (
@@ -74,19 +57,10 @@ function CardList() {
       </div>
 
       <EditModal>
-        <h2>카드 편집</h2>
-        {cardList[focusCardIndex] && <Card {...cardList[focusCardIndex]} />}
-        <div className="button-container flex">
-          <Button type="primary" width="50%" onClick={onClickEditButton}>
-            편집
-          </Button>
-          <Button type="warning" width="50%" onClick={onClickDeleteButton}>
-            삭제
-          </Button>
-        </div>
-        <Button width="100%" onClick={handleEditModalClose}>
-          닫기
-        </Button>
+        <CardEditConfirm
+          focusCardIndex={focusCardIndex}
+          onClickCloseButton={handleEditModalClose}
+        />
       </EditModal>
     </>
   );
